# order-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task",
        "version": "1"
    },
    "paths": {
...
        "/orders/web/v1/orders": {
            "get": {
                "tags": [
                    "web-orders"
                ],
                "operationId": "caf04f4b0bd10f0ebb8db6d384d4b641",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "web-orders"
                ],
                "operationId": "c45b39d158349c8942f0cbb303ea61a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_id": {
                                        "description": "Address",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "Product",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "count": {
                                                    "description": "Count",
                                                    "type": "integer",
                                                    "example": "5"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/orders/web/v1/orders/{id}": {
            "get": {
                "tags": [
                    "web-orders"
                ],
                "operationId": "359efe02a20185a4b5b0d33992ae958b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Description",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Price",
                                            "type": "decimal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "web-orders"
                ],
                "operationId": "0d463dc8fbe4c93f72b4e89677d6d8c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "web-orders"
                ],
                "operationId": "eb81cb88aa274042297fbfa614dcfc43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_id": {
                                        "description": "Address",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "Product",
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "count": {
                                                    "description": "Count",
                                                    "type": "integer",
                                                    "example": "5"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
...
}